+++ tests/time_integration.rs
@@ -2,7 +2,7 @@
 
 use humfmt::{
     time::{self as humtime, TimeHumanize},
-    DurationConversionError, DurationOptions,
+    AgoOptions, DurationConversionError,
 };
 
 #[test]
@@ -32,7 +32,7 @@ fn rejects_negative_time_duration() {
 #[test]
 fn supports_custom_options_for_time_duration() {
     let delta = ::time::Duration::milliseconds(1500);
-    let out = humtime::ago_with(delta, DurationOptions::new().long_units()).unwrap();
+    let out = humtime::ago_with(delta, AgoOptions::new().long_units()).unwrap();
     assert_eq!(out.to_string(), "1 second 500 milliseconds ago");
 }
 
@@ -51,7 +51,7 @@ fn supports_ago_since_with_long_options() {
     let then = ::time::OffsetDateTime::from_unix_timestamp(0).unwrap();
     let now = ::time::OffsetDateTime::from_unix_timestamp(3665).unwrap();
     let out =
-        humtime::ago_since_with(then, now, DurationOptions::new().long_units().max_units(3)).unwrap();
+        humtime::ago_since_with(then, now, AgoOptions::new().long_units().max_units(3)).unwrap();
 
     assert_eq!(out.to_string(), "1 hour 1 minute 5 seconds ago");
 }